+++ AllCropsAllSeasons/Properties/AssemblyInfo.cs
@@ -1,8 +1,5 @@
 ﻿using System.Reflection;
-using System.Runtime.InteropServices;
 
 [assembly: AssemblyTitle("All Crops All Seasons")]
-[assembly: AssemblyDescription("")]
-[assembly: Guid("32464438-69f8-4f22-8f3a-5a5a1f30c563")]
-[assembly: AssemblyVersion("1.4.0.0")]
-[assembly: AssemblyFileVersion("1.4.0.0")]
+[assembly: AssemblyVersion("1.5.0")]
+[assembly: AssemblyFileVersion("1.5.0")]