+++ TimeSpeed/Framework/KeysConfig.cs
@@ -1,4 +1,4 @@
-using Microsoft.Xna.Framework.Input;
+using StardewModdingAPI;
 
 namespace TimeSpeed.Framework
 {
@@ -9,15 +9,15 @@ namespace TimeSpeed.Framework
         ** Accessors
         *********/
         /// <summary>Freeze or unfreeze time. Freezing time will stay in effect until you unfreeze it; unfreezing time will stay in effect until you enter a new location with time settings.</summary>
-        public Keys? FreezeTime { get; set; } = Keys.N;
+        public SButton? FreezeTime { get; set; } = SButton.N;
 
         /// <summary>Slow down time by one second per 10-game-minutes. Combine with Control to increase by 100 seconds, Shift to increase by 10 seconds, or Alt to increase by 0.1 seconds.</summary>
-        public Keys? IncreaseTickInterval { get; set; } = Keys.OemPeriod;
+        public SButton? IncreaseTickInterval { get; set; } = SButton.OemPeriod;
 
         /// <summary>Speed up time by one second per 10-game-minutes. Combine with Control to decrease by 100 seconds, Shift to decrease by 10 seconds, or Alt to decrease by 0.1 seconds.</summary>
-        public Keys? DecreaseTickInterval { get; set; } = Keys.OemComma;
+        public SButton? DecreaseTickInterval { get; set; } = SButton.OemComma;
 
         /// <summary>Reload all values from the config file and apply them immediately. Time will stay frozen if it was frozen via hotkey.</summary>
-        public Keys? ReloadConfig { get; set; } = Keys.B;
+        public SButton? ReloadConfig { get; set; } = SButton.B;
     }
 }